move constants for api functions
